tf2xception/requirements.txt
tensorflow-cpu==2.5.0
pymilvus-orm==2.0.0.rc2
